pico/requirements.txt
pyserial  # required by scripts/readserial.py
